backend/requirements_mac.txt
anyio==3.5.0
asgiref==3.5.0
astroid==2.9.3
asttokens==2.0.5
backcall==0.2.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
clique==2.0.0
colorama==0.4.4
cycler==0.11.0
debugpy==1.5.1
decorator==5.1.1
entrypoints==0.4
executing==0.8.3
fastapi==0.75.0
fonttools==4.32.0
h11==0.13.0
idna==3.3
imageio==2.16.2
ipykernel==6.9.1
ipython==8.1.1
isort==5.10.1
jedi==0.18.1
jupyter-client==7.1.2
jupyter-core==4.9.2
kiwisolver==1.4.2
lazy-object-proxy==1.7.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
mccabe==0.6.1
nest-asyncio==1.5.4
numpy==1.22.3
packaging==21.3
parse==1.19.0
parso==0.8.3
pickleshare==0.7.5
Pillow==9.1.0
platformdirs==2.5.1
prompt-toolkit==3.0.28
pure-eval==0.2.2
pydantic==1.9.0
Pygments==2.11.2
pylint==2.12.2
pyparsing==3.0.8
python-dateutil==2.8.2
PyYAML==6.0
pyzmq==22.3.0
requests==2.27.1
six==1.16.0
sniffio==1.2.0
stack-data==0.2.0
starlette==0.17.1
toml==0.10.2
tornado==6.1
traitlets==5.1.1
typing_extensions==4.1.1
urllib3==1.26.9
uvicorn==0.17.5
wcwidth==0.2.5
wrapt==1.13.3
python-socketio-5.7.1
